refactor(s3_backup): extract S3 backup logic to reusable concern and helpers
- Moved backup-related private methods from S3BackupService to a new S3BackupOperations concern.
- Created S3RakeHelpers module to share S3 validation and error handling logic for rake tasks.
- Updated s3:backup rake tasks to use S3RakeHelpers instead of duplicating code.
- This improves code reuse and maintainability by centralizing backup operations and S3 helpers.
