+++ lib/tasks/s3_backup.rake
@@ -3,86 +3,12 @@
 require "fileutils"
 require "zlib"
 require "rubygems/package"
+require_relative "../../lib/s3_rake_helpers"
 
 namespace :s3 do
   namespace :backup do
-    # Helper methods for S3 operations in rake tasks
-    def ensure_s3_enabled
-      return if ENV["USE_S3_STORAGE"] == "true"
-
-      puts "❌ S3 storage is not enabled. Set USE_S3_STORAGE=true in your .env file"
-      exit 1
-    end
-
-    def validate_s3_config
-      required_vars = %w[S3_ENDPOINT S3_ACCESS_KEY_ID S3_SECRET_ACCESS_KEY S3_BUCKET]
-      missing_vars = required_vars.select { |var| ENV[var].blank? }
-
-      if missing_vars.any?
-        error_msg = "Missing required S3 environment variables: #{missing_vars.join(", ")}"
-        puts "❌ #{error_msg}"
-
-        Sentry.capture_message(error_msg, level: "error", extra: {
-          missing_vars: missing_vars,
-          task: "s3:backup",
-          environment: Rails.env
-        })
-
-        exit 1
-      end
-    end
-
-    def get_s3_service
-      service = ActiveStorage::Blob.service
-
-      unless service.is_a?(ActiveStorage::Service::S3Service)
-        puts "❌ Active Storage is not configured to use S3. Current service: #{service.class.name}"
-        exit 1
-      end
-
-      service
-    end
-
-    def handle_s3_errors
-      yield
-    rescue Aws::S3::Errors::ServiceError => e
-      puts "\n❌ S3 Error: #{e.message}"
-      Sentry.capture_exception(e)
-      exit 1
-    rescue => e
-      puts "\n❌ Unexpected error: #{e.message}"
-      Sentry.capture_exception(e)
-      exit 1
-    end
-    def backup_dir = "db_backups"
-
-    def temp_dir
-      unless Rails.root
-        error_msg = "Rails.root is nil - cannot determine temp directory"
-        Sentry.capture_message(error_msg, level: "error")
-        raise error_msg
-      end
-      Rails.root.join("tmp/backups")
-    end
-
-    def database_path
-      db_config = Rails.configuration.database_configuration[Rails.env]
-      unless db_config && db_config["database"]
-        error_msg = "Database configuration missing for #{Rails.env} environment"
-        Sentry.capture_message(error_msg, level: "error", extra: {
-          rails_env: Rails.env,
-          db_config: db_config
-        })
-        raise error_msg
-      end
-
-      path = db_config["database"]
-      # Handle relative paths
-      path = Rails.root.join(path) unless path.start_with?("/")
-      path
-    end
-
-    def backup_retention_days = 60
+    # Include the shared S3 rake helpers
+    extend S3RakeHelpers
 
     desc "Backup SQLite database to S3 (via job queue)"
     task database: :environment do
@@ -251,52 +177,6 @@ namespace :s3 do
           FileUtils.rm_f(temp_backup_path)
         end
       end
-    end
-
-    private
-
-    def create_tar_gz(source_path, dest_path)
-      # Check if tar command is available
-      unless system("which tar > /dev/null 2>&1")
-        error_msg = "tar command not found - required for backup compression"
-        Sentry.capture_message(error_msg, level: "error", extra: {
-          environment: Rails.env,
-          path: ENV["PATH"]
-        })
-        raise error_msg
-      end
-
-      # Use system tar command for reliable compression
-      dir_name = File.dirname(source_path)
-      base_name = File.basename(source_path)
-      tar_command = "tar -czf #{dest_path} -C #{dir_name} #{base_name}"
-
-      unless system(tar_command, exception: true)
-        error_msg = "Failed to create tar archive"
-        Sentry.capture_message(error_msg, level: "error", extra: {
-          command: tar_command,
-          source_path: source_path,
-          dest_path: dest_path
-        })
-        raise error_msg
-      end
-    end
-
-    def cleanup_old_backups(service)
-      bucket = service.send(:bucket)
-      cutoff_date = Time.current - backup_retention_days.days
-      deleted_count = 0
-
-      bucket.objects(prefix: "#{backup_dir}/").each do |object|
-        next unless object.key.match?(/database-\d{4}-\d{2}-\d{2}\.tar\.gz$/)
-
-        if object.last_modified < cutoff_date
-          service.delete(object.key)
-          deleted_count += 1
-        end
-      end
-
-      print "(deleted #{deleted_count} old backups) " if deleted_count.positive?
     end
   end
 end
